Introduce spotbugs (#669)

+++ bboxdb-server/src/main/java/org/bboxdb/distribution/allocator/CPUCoreUtilizationAllocator.java
@@ -52,7 +52,7 @@ public class CPUCoreUtilizationAllocator extends AbstractUtilizationAllocator {
 			return 0;
 		}
 		
-		return distributedInstance.getCpuCores() / count;
+		return (double) distributedInstance.getCpuCores() / count;
 	}
 
 	@Override
